Improve change point detection reliability (#10)
Improve change point detection reliability

We noticed that EDivisive algorithm is unreliable
if run on a very long series of data with many
cancelling changes. If a regression is present for
a short time, the time may be too short in proportion
to the length of the entire signal, it may be missed.

This commit modifies the way how EDivisive algorithm
is applied to our data. Instead of running it on
a full series, we apply it to a shorter window
moving over the data. This way the amount of data
per each invocation of EDivisive is smaller and
it can see fewer change points at a time
making it more reliable.

A nice bonus side effect is that this
reduces the time complexity from O(n^2) to O(n k^2)
where k window size, n length of the series.