Remove change points with small change magnitude
This commit adds a change-point post-processing step.
In that step, weak change-points are removed based on
their p-value and change magnitude. This allows us
to bump up the p-value limit in the e-divisive
part of the algorithm, potentially reducing the risk
of missing some points. In Edivisive, change points
to found later typically increase the p-value of
change-points found earlier; hence the algorithm can
miss some good points.

Fixes #25