Remove dependency on fallout (#60)
Remove dependency on fallout

The biggest change introduced in this commit is moving the test
configuration into hunter.yaml file. This way the results of tests
launched from systems other than Fallout can be analyzed, as long as
they send data to Graphite. Previously contacting Fallout was needed to
fetch the export prefix of the test, but now all the required
information is located in the config file.

Ad-hoc (command-line) test configuration has been removed. A test must
be defined in the config file before it can be analyzed. This simplifies
hunter a lot and although it may initially feel a bit less
user-friendly, having a single way of passing test metadata to hunter
makes it easier to reason about. If this turns out to be too limiting,
we may add a way to submit config files (or override config file
properties) from the command line in the future.

Unifying configuration allowed the `analyze` and `bulk-analyze`
commands to be merged into a single `analyze` command which accepts a
list of test names or test group names.

Reduce the length of the target query in Graphite