Extract change point analysis results as AnalyzedSeries class
Returning a vector of change points is not good enough, because
in some cases we want to process change points one metric at a time,
so we want them to be grouped by metric name, not time.

Additionally, this makes code for comparing two time series
or comparing different time points in a single series cleaner,
because the client code doesn't need to compute and store
change points in additional variables.