New command: hunter regressions
The regressions command compares the performance of the
end of the selected range with the performance at the beginning
of the selected range.

E.g. hunter regressions <test> --since '10 weeks ago' compares
the current average performance level with the average performance about
10 weeks ago. If current performance is worse significantly (in terms
of statistical significance) than the baseline performance on any of
the metrics, it is reported to stdout.

Fixes #13