Finding regressions on a feature branch
A new configuration property has been added to graphite
tests: `branch_prefix`.

If the user specifies `--branch <name>` argument, the data
are fetched from the prefix set in `branch_prefix`.
This works for both `analyze` and `regressions` commands.

The `regressions` command has been modified to compare
performance of the branch selected with `--branch` to the
baseline taken from the main branch, located by the main prefix.